requirements.txt
scipy>=1.1.0
numpy>=1.17.3
scikit-image>=0.16.2
OWSLib>=0.19.1
pandas>=0.23.4
rasterio>=1.1.3
Pillow>=5.2.0
matplotlib>=3.0.3
libpysal>=4.7.0
networkx>=3.1
pyyaml